Implement `ADD A, $imm8` and simplify code
We don't have to use `std::copy()` for the copying of data from the
cartridge file to our vector, as long as we use a `basic_ifstream`
specified as an `unsigned char`. From there, we can just use the
`istreambuf_iterator` directly as an argument to the vector, populating
it.